+++ iOS/second-hand/second-hand/Chatting/ChattingViewController.swift
@@ -15,12 +15,7 @@ class ChattingViewController: UIViewController {
         let attributes = [NSAttributedString.Key.font: UIFont.systemFont(ofSize: 17, weight: .bold)]
         self.navigationController?.navigationBar.titleTextAttributes = attributes
         self.navigationItem.title = "채팅"
-        setNavigationBarBottomBorder()
+
     }
     
-    func setNavigationBarBottomBorder() {
-            let borderView = UIView(frame: CGRect(x: 0, y: navigationController?.navigationBar.frame.height ?? 0 - 1, width: navigationController?.navigationBar.frame.width ?? 0, height: 1))
-            borderView.backgroundColor = .lightGray
-            navigationController?.navigationBar.addSubview(borderView)
-        }
 }
